crunevo_cascaron/crunevo/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Migrate
python-dotenv
psycopg2-binary
